Refactor slightly the way that function and global aliases are computed so they get cached in the `LLVMContext` instead of being computed on demand.
